Timer Reset bug lifeMinistryTimer.js
### Fixed the bug in the timer-reset button. The issue was in the resetTimer method in the lifeMinistryTimer.js file.

### The Problem
The bug occurred because the resetTimer method was stopping the timer regardless of which part's timer was being reset. When you reset the timer for meeting part A while meeting part B was actively timing, it would stop the timer for part B as well.

### The Solution
Modified the resetTimer method to only stop the timer if the part being reset is the currently active part.

### With this change, resetting a non-active part's timer will:
Reset that part's elapsed time to zero
Update the display
NOT stop the timer for the currently active part

This ensures that the reset button on specific parts will only affect the timer for that specific part, as expected.